gop_scraper/requirements.txt

playwright==1.45.0
python-dotenv==1.0.1
pandas==2.2.2
